johnson/requirements.txt
amqp==5.1.1
asgiref==3.5.1
billiard==3.6.4.0
celery==5.2.7
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
cycler==0.11.0
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.6
distributor==2.4.2
Django==4.0.4
django-appconf==1.0.5
django-auditlog==2.0.0
django-bmemcached==0.3.0
django-bootstrap-datepicker-plus==4.0.0
django-bootstrap-daterangepicker==1.0.6
django-bulma==0.8.3
django-celery-beat==2.3.0
django-crispy-forms==1.14.0
django-discover-runner==1.0
django-excel==0.0.10
django-filter==22.1
django-gst-field==0.0.3
django-import-export==2.8.0
django-jazzmin==2.6.0
django-mathfilters==1.0.0
django-multiselectfield==0.1.12
django-phone-field==1.8.1
django-report-builder==6.4.2
django-role-permissions==3.1.1
django-roles-access==0.9.4
django-select2==8.0.0
django-timezone-field==5.0
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
dnspython==1.15.0
dnspython3==1.15.0
et-xmlfile==1.1.0
Faker==13.12.0
filelock==3.8.0
fonttools==4.33.3
idna==3.3
jdcal==1.4.1
Jinja2==3.1.2
kiwisolver==1.4.3
kombu==5.2.4
lml==0.1.0
MarkupPy==1.14
MarkupSafe==2.1.1
matplotlib==3.5.2
numpy==1.22.4
odfpy==1.4.1
openpyxl==2.6.0
packaging==21.3
pandas==1.4.2
Pillow==9.1.1
platformdirs==2.5.2
postgres==4.0
prompt-toolkit==3.0.30
psycopg2==2.9.3
psycopg2-binary==2.9.3
psycopg2-pool==1.1
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-webio==0.1.4
pyparsing==3.0.9
python-binary-memcached==0.31.1
python-crontab==2.6.0
python-dateutil==2.8.2
python-memcached==1.59
pytz==2022.1
PyYAML==6.0
reportlab==3.6.10
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tablib==3.2.1
texttable==1.6.4
tzdata==2022.1
uhashring==2.2
urllib3==1.26.9
vine==5.0.0
virtualenv==20.16.5
virtualenvwrapper-win==1.2.7
wcwidth==0.2.5
xlrd==2.0.1
XlsxWriter==3.0.3
xlutils==2.0.0
xlwt==1.3.0
